Fixing demo.py to work with numpy image.
Previous commits by others have reworked the image
handling to avoid PIL images and PNG splitters and instead
use numpy arrays of pixels.

This change amends demo.py so that it respects that change
and it starts to display an image again. And it's quick too!

This will break compatibility with AR.Drone 1, but by now it's
safe to assume that was broken ages ago in various ways.